pyproject.toml
[tool.poetry]
name = "lannister_bot"
version = "0.1.0"
description = "poggers bot that will feed starving children in africa"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
slack-sdk = "^3.18.1"
psycopg2-binary = "^2.9.3"
drf-spectacular = "^0.23.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
coveralls = "^3.3.1"
pytest-mock = "^3.8.2"
flake8 = "^4.0.1"
black = "^22.6.0"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
